+++ src/autovala.vala
@@ -20,11 +20,11 @@ using GLib;
 using Gee;
 using Posix;
 
-//project version = 0.99.46
+//project version = 0.99.47
 
 void help() {
 
-	GLib.stdout.printf(_("Autovala. Usage:\n\tautovala help: shows this help\n\tautovala version: shows the current version\n\tautovala init project_name: initializates a new Vala CMake project and creates an initial project file\n\tautovala refresh: tries to guess the type for each file in the folders and adds them to the project file\n\tautovala cmake: creates the CMake files from the project file\n\tautovala update: the same than 'refresh'+'cmake'\n\tautovala po: updates translatable strings\n\tautovala clear: removes the automatic parts in the project file, leaving only the manual ones.\n\tautovala project_files: lists all the files belonging to the project (with paths relative to the project's root). Useful for adding all the files to a versioning system like git, bazaar or subversion\n\tautovala deb: creates the 'debian' folder for packaging the project as a .deb package\n\tautovala rpm: creates the 'rpmbuild' folder for packaging the project as a .rpm package\n\tautovala pacman: creates a package for PACMAN package manager\n\tautovala valama: exports the project to a VALAMA project file\n\n"));
+	GLib.stdout.printf(_("Autovala. Usage:\n\tautovala help: shows this help\n\tautovala version: shows the current version\n\tautovala init project_name: initializates a new Vala CMake project and creates an initial project file\n\tautovala refresh: tries to guess the type for each file in the folders and adds them to the project file\n\tautovala cmake: creates the CMake files from the project file\n\tautovala update: the same than 'refresh'+'cmake'\n\tautovala po: updates translatable strings\n\tautovala clear: removes the automatic parts in the project file, leaving only the manual ones.\n\tautovala project_files: lists all the files belonging to the project (with paths relative to the project's root). Useful for adding all the files to a versioning system like git, bazaar or subversion\n\tautovala deb: creates the 'debian' folder for packaging the project as a .deb package\n\tautovala rpm: creates the 'rpmbuild' folder for packaging the project as a .rpm package\n\tautovala pacman: creates a package for PACMAN package manager\n\tautovala valama: exports the project to a VALAMA project file\n\tautovala external owner_id: shows the external data of the specified owner\n\n"));
 }
 
 int main(string[] argv) {
@@ -230,6 +230,17 @@ int main(string[] argv) {
 		}
 		generate_valamang(data);
 		GLib.stderr.printf(_("Done\n"));
+		break;
+	case "external":
+		if (argv.length!=3) {
+			help();
+			return -1;
+		}
+		var gen = new AutoVala.ManageProject();
+		var project = gen.get_binaries_list(null,argv[2]);
+		foreach(var data in project.external) {
+			GLib.stdout.printf("%s: %s\n",argv[2],data);
+		}
 		break;
 	default:
 		help();