+++ src/autovalaLib/elementDesktop.vala
@@ -40,7 +40,7 @@ namespace AutoVala {
 			return error;
 		}
 
-		public override bool configureLine(string line, bool automatic, string? condition, bool invertCondition, int lineNumber) {
+		public override bool configureLine(string line, bool automatic, string? condition, bool invertCondition, int lineNumber, string[]? comments) {
 
 			// The line starts with 'binary: '
 			if (line.has_prefix("desktop: ")) {
@@ -55,6 +55,7 @@ namespace AutoVala {
 				return true;
 			}
 			var data=line.substring(2+this.command.length).strip();
+			this.comments = comments;
 			return this.configureElement(data,null,null,automatic,condition,invertCondition);
 		}